Start integrating SPDM into sled-agent (#407)
Start integrating SPDM into sled-agent

This code only represents the negotiation phase of the protocol.

A follow up commit will contain the responder identification (digests + cert
retrieval) code, as well as challenge response, which is the first part of the
protocol used to begin securing a SPDM channel. These are the parts of the SPDM
implementation that exist so far. More will be added over time.

It's expected that `run` for each of a requester and responder will be used by a
caller establishing a secure channel. That usage will come in a follow up
commit. It's likely that when the SPDM protocol is complete a secure channel
will be returned for writing to. Until then, callers will pretend that run sets
up a fully secure channel, and then use the existing TCP stream when run
returns.


Co-authored-by: Sean Klein <sean@oxide.computer>