Add code for creating, splitting, combining a RackSecret (#429)
Add code for creating, splitting, combining a RackSecret

A new `RackSecret` type used to implement the "Trust Quorum" functionality of
the rack is provided in this commit. It uses Shamir secret sharing with Feldman
verification. Feldman verification was chosen for simplicity as it doesn't
require creating blinding shares and distributing those along with the actual
shares as with Pederson verification. We can always change this later.

The NIST P-256 curve was chosen somewhat arbitrarily, although it is a popular
curve and works with Shamir secret sharing without any extra mechanisms required
by some other curves.

A very simple protocol will be created to use this functionality over
SPDM channels.