Remove all the ugly trust quorum config and related generation code
Replace this with a `ShareDistribution` type that will be generated by a
deployment system coming in a follow up commit. In the newer code only
individual shares for a given sled are handed out, so the sled-agent can't cheat
and unlock itself from reading a file.

Also cleanup all the other stuff related to Sean's review.