Fixes racy oximeter integration test
Fixes a race in the oximeter integration test that verifies behavior
when the producer or collector dies and comes back. The race can occur
when the collector scrapes a sample, then the producer is dropped. At
this point, the test gets the new timeseries, and asserts it's the same
as the old. That's not quite right, since it's possible for the new
sample to have been produced immediately before the producer is dropped.
Then it appears on the next query, and the assertion is tripped.

The test is modified to be slightly weaker. It asserts that the
timeseries match up to the number of samples previously collected, and
that any new samples have timestamps before the time the producer is
dropped.