Updates default VPC Subnet behavior
This improves the API and database internals around VPC Subnets,
specifically the how we create the default VPC Subnet for a VPC, and how
IP address ranges are specified by clients or generated internally.

It also adds a query that checks for overlapping address ranges, and
fails the insert of a new VPC Subnet if an existing subnet in the same
VPC has overlapping IP address ranges, either IPv4 or IPv6.