Leave one Instance as authoritative in sending state updates to Nexus. (#667)
For every Instance sled agent manages, it spawns a task to poll for state updates from propolis. After updating its in-memory state, it notifies Nexus to update the DB representation.

In the Live Migration case, we're in a transitory state wherein there will be two distinct Instance objects (possibly across 2 different Sled Agents). We update Nexus here to only write the changes from the destination SA to the DB.

Along with that, this pulls in the latest propolis side changes. We're still lacking actually transfering over all the necessary state but a lot more scaffolding is coming together.