Update with crucible agent changes
The crucible agent now requires clients to specify that a region should
be created with the expectation that an upstairs will connect encrypted
or not, along with X509 information for launching the corresponding
downstairs instances with a TLS acceptor.

The region in the DB now has the "encrypted" field, but the DiskCreate
struct accepts an encryption key UUID. It's still a TODO to create a
key, associate it with a region set, and launch any Upstairs
appropriately.