use a single updates base URL
we intend to have an updates repository that customers can mirror on
their internal networks in case they make the reasonable decision to not
plug their racks into the internet.

also: at some point we may choose to pivot the metadata URL used to
"hide" updates from earlier versions of nexus, in case we don't invent
another mechanism to do that. that is, we would publish one final nexus
update to the update metadata in /metadata/, and that update would
change nexus to use /metadata-new/. this allows us to force a two-stage
update if necessary.

or, perhaps, we might have other update channels that can be configured
on the rack. a user may want to say "update to beta". having the entry
point be a single URL allows us to build these abstractions without
telling the user to change the URL in a specific way or requiring
specific a specific mirroring setup.