[v2] TUF integration in Nexus + update artifact fetching by sled-agent (#717)
* WIP Nexus download endpoint

* WIP writing files

* free function now method

* sorting out errs

* Add tests, fix bugs

* I guess compiling code is better than the alternative

* I EXPECTORATEd that my JSON would be more up-to-date

* initial work for using the tough client in nexus

* run tough client in tokio::task::spawn_blocking

nexus is async, tough is not async, but tough uses reqwest in blocking
mode which is async.

https://docs.rs/reqwest/0.11.7/reqwest/blocking/index.html:

> Conversely, the functionality in `reqwest::blocking` must not be
> executed within an async runtime, or it will panic when attempting to
> block.

moves the relevant code out to updates.rs, since the function can't
borrow self due to lifetime constraints.

* allow an unconfigured updates system

* woo license header

* fix hardcoded base URLs to use localhost, for now

* add smf/nexus/root.json to .gitignore

* rename columns to make more sense at first glance

* keep table in sync with artifacts.json

* make the tests happy

* add /updates/refresh to oxapi_demo

* fix examples/config-file.toml

* test [updates] in config works

* remove dead code

* .unwrap() on JoinError from spawn_blocking

... since We only get a JoinError if the task we're waiting on panics.

* wire download_artifact up to the database

* tell all sleds to apply all updates

* fixup! wire download_artifact up to the database

* actually verify the target after download

* Keep merging

* use ..UpdatedNoContent instead of ..Ok<()>

this fixes problematic generation of an OpenAPI interface, which in turn
fixes generation of non-compiling code where `()` becomes an enum with
no variants.

* end-to-end updates test

* move UpdateArtifactKind into common

* use a single updates base URL

we intend to have an updates repository that customers can mirror on
their internal networks in case they make the reasonable decision to not
plug their racks into the internet.

also: at some point we may choose to pivot the metadata URL used to
"hide" updates from earlier versions of nexus, in case we don't invent
another mechanism to do that. that is, we would publish one final nexus
update to the update metadata in /metadata/, and that update would
change nexus to use /metadata-new/. this allows us to force a two-stage
update if necessary.

or, perhaps, we might have other update channels that can be configured
on the rack. a user may want to say "update to beta". having the entry
point be a single URL allows us to build these abstractions without
telling the user to change the URL in a specific way or requiring
specific a specific mirroring setup.

* clean up unnecessary ResourceType variants

* refactor to require full artifact descriptions

* work around dropshot Response<Body> issue

* undo this change

* combine updates integ tests into one module

* remove errant fixme

* comment nit

Co-authored-by: Sean Klein <sean@oxide.computer>