Read-only instance serial console API endpoint in Nexus.
This first-pass implementation is simply a GET with parameters for
byte offset since boot and length, with the intent of providing something
simple for the web frontend to use. It's implemented by cacheing both the
first MiB and most recent MiB of serial console output in the sled-agent,
though this detail is not important in itself (i.e. it could just as well
have the meat of the implementation moved to propolis-server.)