Remove the use of default feature for vsss_rs (#794)
Remove the use of default feature for vsss_rs

The use of the curve25519 feature was causing build failures after running
cargo update. This was a result of a bumping the vsss_rs dependency
`curve25519-dalek` from "4.0.0-pre.1" to "4.0.0.-pre.2". This used a transitive
dependency of "digest" which has a backwards incompatible change from 0.9 to
0.10 and was triggering failure of the ed25519-dalek build.

Since we aren't using either of ed25519 or secp256k1, we just remove the use of
the default feature altogether. This fixes the build.