Fix-up instance creation saga so it actually works
I had an incorrect mental model of how distributed sagas work and this
caused problems - disks would not be detached when instance creation
failed.

I tried to correct the code in a slightly different way than was done
for NICs. Instead of pre-allocating IDs and having a previous saga node
delete based on these IDs, I chose to split the attach disks to instance
node into 8, each with its own undo.

This is now tested by trying to attach 8 disks, one of which is marked
as faulted. The whole saga should unwind and the test checks for
semantic equivalence.

Added some tests for the new arbitrary 8 disk limit. This required
bumping the request max body size because of the larger instance ensure
bodies.