Block size must be given during disk create
Block size of disks must be variable and declared during disk creation.
In other instances, it can be inherited (for example, creating a disk
from a snapshot will set the block size to whatever the snapshot's was).
Store this in the disks table.

Also, prepare for creating disks from images and snapshots by adding
the appropriate columns to the disks table.

Externally, the API should be returning names, not IDs, but this commit
doesn't tackle that. At minimum this will require Nexus to convert a
db::model::Disk into an external::Disk by looking up ids and filling in
the appropriate names (instead of a simple From impl). I'll create an
issue to track this work.