Implement global images
Fill in image creation code, adding a few fields to the image from RFD
4 (version and digest specifically). Also add block_size due to the
constraint that block_size must be present during any volume creation.

Store volume construction requests created from the parameters in
global_image_create. This commit does not implement global_image_delete
because that involves much work to figure out how to safely do this (are
any disks using it? are other volumes using this as a sub volume?).

Both "Image" and "GlobalImage" are required to separate authz policy, so
this commit separates the two. Project images will be implemented in
another PR.

Implement authz for global images. GlobalImagesList is a dummy authz
resource that was added to have a place to check list_children
permissions for.