[package] Remove sled-agent config during install/uninstall (#888)
## Background

The sled agent makes use of a directory on local storage to store a variety of configuration info, including:

- What services am I managing? What services should I start on reboot? How should I boot them?
- What storage am I managing? Which datasets are managed by Cockroach, Clickhouse, Crucible? How should they be initialized?

This info is pretty handy, and is almost entirely self-managed by the sled itself. *However*, Omicron is still under
development, so this format isn't really stable, because our configs aren't stable.

In the face of changing configs, we currently throw errors:

- https://github.com/oxidecomputer/omicron/blob/65e0804195e8d365b56418d691bfb00de6ed79b7/sled-agent/src/rack_setup/service.rs#L110-L122
- https://github.com/oxidecomputer/omicron/blob/65e0804195e8d365b56418d691bfb00de6ed79b7/sled-agent/src/services.rs#L222-L227

This can detect failures, but basically requires developers to manually remove these config directories.

## What does this PR do

It removes the configuration directory - effectively wiping the Sled - on install/uninstall.
Given that the install process is intended to be a "wipe, restart the world" process - as opposed the upgrade flow - this
seemed like a reasonable way to improve the developer experience.