MGS: Change serial console API to use websockets (#883)
Clients can attach (if there is not another attached connection), and
data they send will be shuffled to and from the SP via UDP. At the
moment buffering that data into reasonable sizes is the responsibility
of the client; I'm not sure that's right, but it seems okay for now. The
new `gateway-cli` binary buffers by line by default, or in 500ms chunks
if the user requests raw mode.