prevent accidentally destroying shared machines
Some of the scripts in tools/ are increasingly destructive, and can
be deleterious when executed on a shared build machine where the
user has appropriate privileges.  To avoid issues, we will look for
a marker file that should prevent us from doing anything (especially
as root) to mutate the machine.  To marker a build machine as
off-limits:

    # mkdir -p /etc/opt/oxide
    # touch /etc/opt/oxide/NO_INSTALL