Support for SAML as a Silo IdP, part 1
Add the db schemas, models, and some endpoints to support configuring a
SAML IdP for a Silo. Enough functionality is here to support the first
step of SP-initiated SAML login flow. More work is required to support
receiving the SAML IdP's response, and actually creating and logging in
the user.

Two tables were added here: one that relates a silo to a list of typed
identity providers, and one for saml configuration. Future work will add
"local" and "ldap" identity provider support.

Table column order was corrected - I didn't see any bugs from this but
am aware that the bug potential exists.

Currently, the external-authenticator role only applies to the
default silo, not any new ones. Making that opctx work with any silo
will unlock some more testing.

This commit also moves the unauthorized_coverage code into the
unauthorized function, and sets up both hardcoded and new non-hardcoded
data. When global images support was added, a hardcoded port was used
for the httptest server and it was a problem to find an unused port, but
allowing for non-hardcoded data means any port can be used.