internal-dns: NXDOMAIN/SERVFAIL where appropriate (#1054)
* internal-dns: NXDOMAIN/SERVFAIL where appropriate

Currently, when the internal DNS server encounters an error or has no
records to serve, it will just return a DNS response with no resource
records. This causes issues for resolvers that expect:

- A SERVFAIL for records the DNS server cannot resolve.
- An NXDOMAIN for records the DNS server asserts do not exist.

This patch updates the internal DNS server to conform to those
expectations.

This required adding a zone configuration option to the internal DNS
server that identifies what domain it's presiding over. Any records that
do not fall within that domain will result in a SERVFAIL, causing
resolvers to look for other DNS servers for answers. This particular
failure mode falls under the SERVFAIL DNS Error Code 7 [1] "No Reachable
Authority" as the internal DNS server does not take responsibility for
recursive resolution.

[1] https://tools.ietf.org/id/draft-ietf-dnsop-extended-error-05.html#rfc.section.4.2.7

* internal-dns: send SERVFAIL on deserialize error