deserialize_with public cert and private keys
Use serde's deserialize_with to validate public cert and private keys,
causing failure if the DerEncodedKeyPair isn't valid. The visitor used
here is a fn from `String -> Result<String>` which allows for early
validation, and afterwards any code operating on DerEncodedKeyPair can
be sure it contains valid data.

Unfortunately, the deserialize_with could not be a fn from `String ->
an openssl::type` because the openssl types do not derive JsonSchema.

Previously some of this validation was in SiloSamlIdentityProvider's
validate, and that has been removed in this commit.