Add authn::SiloSamlIdentityProvider
Instead of putting an impl for the model::SiloSamlIdentityProvider type
into the authn subsystem, make a new authn type, and convert the model
type into that.

Move the logic that looks up SiloIdentityProviderType from nexus.rs to
authn silo subsystem.

Move validate function logic into a new TryFrom impl for the authn
SiloSamlIdentityProvider type. This way it's part of the conversion and
cannot be forgotten about.