Add OPTE uninstallation script
Adds ./tools/uninstall_opte.sh. This can be used to revert a system
which is on an xde-compatible kernel back to stock Helios bits. This
also updates the install_opte.sh script a bit to install exactly the
osnet-incorporation consolidation package that we expect to need in
order to be compatible with xde and OPTE.

Using this, we should be able to switch back and forth between an
xde-compatible and stock Helios kernel with minimal fuss.