Global images should have distribution and version
Global images should have version, according to RFD 4, plus OS
distribution, which is added with this commit. This allows for bucketing
what global images are available to use for booting. This commit
separates the param structs that are used to create global images from
project images.

This commit also adds a new error type called InterregnumError. It's
meant to be returned in cases where different Nexus are operating with
different types. The example in this PR is the difference between the
database enum variants and the external param enum variants. FreeBSD is
provided as an example where there's a database variant but not an
external enum variant. Note we also don't have FreeBSD cloud images
wired up yet for testing.

Note that ImageSource was also changed here to remove `content` from the
serde block, and change the enum variants to use field names. This was
done so that the JSON looked better.