sled-agent initializes sleds on RSS's behalf
Previously RSS itself connected directly to the bootstrap-agent server
on all remote sleds to send initialization requests. This PR changes
that path to have RSS send requests to its local sled-agent (currently
via an in-memory channel, but potentially through something like a Unix
domain socket), and the sled-agent then connects to the bootstrap-agent
servers to actually send the requests to all sleds.

This is done in preparation for wrapping the sled-agent-to-sled-agent
bootstrapping communications in sprockets per RFD 238.