Move sled initialization requests from RSS to bootstrap-agent (#1117)
Previously RSS itself connected directly to the bootstrap-agent server
on all remote sleds to send initialization requests. This PR changes
that path to have RSS send requests to its local bootstrap-agent (currently
via an in-memory channel, but potentially through something like a Unix
domain socket), and the local bootstrap-agent then connects to the remote
bootstrap-agent dropshot servers to actually send the requests to all sleds.

This is done in preparation for wrapping the bootstrap-agent-to-bootstrap-agent
communications in sprockets per RFD 238.