Global images should have distribution and version (#1092)
Global images should have version, according to RFD 4, plus OS distribution,
which is added with this commit. This allows for bucketing what global images
are available to use for booting. This commit separates the param structs that
are used to create global images from project images.

This commit also adds a new error type called TypeVersionMismatch. It's meant to
be returned in cases where different Nexus are operating with different types.
The original PR stored Distribution as an enum, and TypeVersionMismatch was
included to return a specific error case where variant mismatch occured between 
params, database, and views enums. This was removed but TypeVersionMismatch was
left in.

Note that ImageSource was also changed here to remove `content` from the serde
block, and change the enum variants to use field names. This was done so that
the JSON POST body looked better.