Fail if there isn't available space for a disk
Nexus' current region allocation algorithm does not check to see if the
dataset usage goes over the zpool's total size. Add this check, plus
a test to validate it.

The algorithm currently does an all or nothing check: it tries to find
three datasets with available space, and if it cannot then (with this
commit) returns an error - in this commit, the test
`test_disk_backed_by_multiple_regions` fails. When the algorithm changes
to support chunking a disk across multiple dataset triples, that test
will pass. The reason that this was committed too was to make this
explicit to people searching the tests for usage examples.