[internal-dns] Avoid 'picking ports' (#1233)
Fixes #1200

The internal-dns crate includes two servers: a "DNS server" and a "Dropshot server". This change refactors the launching of both to ensure that their ports are queryable, especially under tests conditions where they are now set to "port zero" (let the OS pick for us).

This PR also removes the dependency on the "port-picker" crate.