Authenticated user can list and read global images
Add polar statements so that any authenticated user can list and read
global images.

In order to test this, the set of integration tests that check endpoints
for their authz visibility had to change: `GET /images/` was now
accessible by any authenticated user, even if they had no permissions at
all, meaning new type of visibility was required. But `POST /images` was
protected by authz, meaning separate visibility settings were required
for different HTTP methods. This commit also adds per-method authz
visibility and updates each VerifyEndpoint struct accordingly.

A few of the endpoints in uncovered-authz-endpoints.txt were able to be
added in because of the new visibility variant, but not all.  In order
to test those for authz visibility, more work is required to support
endpoints that redirect (for example) or endpoints with query
parameters.