Connect Crucible stats through Propolis to Oximeter
This commit makes the final step in plumbing metrics genrated in
Crucible Upstairs through Propolis to Oximeter.  Updated versions
of both Crucible and Propolis are required (and they both require
this update).

Removed two fields of CrucibleOpts that are obsolete.

Updates to supply a UUID to the VolumeConstructionRequest.
It is currently just a random UUID, but that is how it was on the
crucible side already, and this change will allow Nexus, when it is
ready, to supply a real UUID. This change is not required for the
metrics, but the change is already in the Crucible side, so it
must be included when we update the Crucible rev.

New git rev and packages for both Propolis and Crucible