buildomat: reshape Helios CI jobs (#1444)
This change splits out the Helios deploy check into a separate job,
"check-omicron-deployment (helios)", which should shorten the critical
path through the build-and-test job where it used to be.  This new job
has a name of the same form as the Actions equivalent on the Linux and
Macintosh side.

In addition, the build and test job should use the new builder-only
prerequisites script; we're not rebooting into the OPTE bits it would
install anyway so this is wasted time.

Finally, we're renaming the build and test job to "build-and-test
(helios)" to align the naming with the Actions jobs.