Add support for Silo groups (#1358)
Add the necessary logic to:

- provision an "admin" silo group when a silo is created, which is
  granted silo admin role.
- after successful authentication, create groups during silo user
  provision if the Silo's provision type is JIT.
- add a group's roles to a user's role set if they're part of that
  group.

Silos now have an optional admin_group_name that is configured at silo
provision time. If this is left out, users will currently have no way to
be granted roles when they first log in. In the future, this may be
selected and groups would be created another way.

SAML identity providers now have an optional group_attribute_name that
configures what attribute represents a group name. Groups can be passed
in multiple attribute values, or in one as a comma separated list.