Compute size_used using query inside transaction
Instead of adding and subtracting dataset's size_used field, compute it
each time by summing up all the sizes of regions that belong to a
dataset. Also sum up all regions belonging to a zpool to validate that
the zpool is not running out of space.

This involved changing the disk related tests to create three zpools,
each with one dataset each, instead of one zpool with three datasets.
Code that was iterating over each dataset now has to iterate over
zpools, then datasets in the zpool.

Updated tests that are now correctly checking for out-of-space.