Remove sanity check from populate-images (#1623)
Let's suppose we have a multi-machine setup:
1. My laptop
2. Folgers (runner)
3. Catacomb (image hosting)

This script is mostly sending requests from machine 1 to machine 2, asking machine 2 to pull images from machine 3.
However, the sanity check at the end checked that machine 1 had access to machine 3. This connectivity may not exist!

In fact, in many demo environments, this connectivity *does not* exist, and the script hangs when doing the sanity check.

To resolve, I'm removing this bit.