Fail if there isn't available space for a disk (#1231)
Nexus' current region allocation algorithm does not check to see if the
dataset usage goes over the zpool's total size. Add this check, plus a
test to validate it.

The algorithm currently does an all or nothing check: it tries to find
three datasets with available space, and if it cannot then (with this
commit) returns an error - in this commit, the test
`test_disk_backed_by_multiple_regions` fails. When the algorithm changes
to support chunking a disk across multiple dataset triples, that test
will pass. The reason that this was committed too was to make this
explicit to people searching the tests for usage examples.

Instead of adding and subtracting dataset's size_used field, compute it
each time by summing up all the sizes of regions that belong to a
dataset. Also sum up all regions belonging to a zpool to validate that
the zpool is not running out of space. This commit also fixes a bug in
regions_hard_delete: previously, that function was only subtracting
size_used for the first dataset in the list of three.

This involved changing the disk related tests to create three zpools,
each with one dataset each, instead of one zpool with three datasets.
Code that was iterating over each dataset now has to iterate over
zpools, then datasets in the zpool.

Updated tests that are now correctly checking for out-of-space.

Because diesel's return type for a summation of INT is Numeric, a
TryFrom impl was added for ByteCount. I couldn't figure out a way to go
from Numeric to i64.  Because of this TryFrom, nexus-db-model now
requires pq-sys (and the rpath fix).