Start implementing bootstore: Replicated storage required for rack unlock (#1656)
Currently RSS is in charge of generating the rack secret and the bootstrap agent is in charge of establishing sprockets sessions and transferring shares required for rack unlock. The shares themselves, and the trust quorum membership is stored in a file on each sled. The major problems with this implementation are that it is not resilient to failure and does not allow reconfiguration of the trust quorum. Additionally, the trust quorum code is somewhat scattered across the sled-agent, and a little hard to follow.

This change introduces the `bootstore`, which will become responsible for all trust quorum protocol and storage on the rack. The bootstore will be instantiated by the boot agent, which will be responsible for establishing sprockets sessions that the trust quorum can operate over. This allows for a readable, testable implementation of the protocol and storage without having to worry about extraneous details related to networking or the SP. We may choose to pull that in later to a higher level module in the bootstore. I intend to utilize property based tests with fault injection to cover the protocol.

This is an early WIP of the bootstore, and lays out the basic structure of the code, while implementing some small bits. The goal is to get this out early before it gets too large to easily review. 

The `Node` is the main entry point to the codebase, although most logic has not been implemented. The `Node` uses the `Db` to query and store trust quorum information. The goal is to build up a fairly general two-phase commit mechanism that can also be used for storing networking information required to setup NTP so that CockroachDb can boot and the rest of the control plane software can run. The two-phase commit is standard, but as always has the problem of any node failing halts the protocol. We will manage this by moving to a new epoch and reconfiguration if this occurs. Of course the devil is in the details, but it's not incredibly complicated. This implementation will supercede what is written in section 5 of RFD 238. I will update that document with the new protocol. One thing that was not considered there was transferring the encrypted (wrapped) root secret to new nodes that don't have the original root secret derived from the epoch 0 rack secret. As part of this, we'll distribute it as part of the 2-phase commit. There are a couple mechanisms with various levels of security, that I won't get into here. The important thing to know for this PR is that I created a table and model for storing the secret. Code to use it is unimplemented so far.

There is also the notion of a [`Coordinator`] that is not yet implemented. The coordinator interacts with the sled-agent to record important information in nexus and get instructions related to trust quorum from nexus. It only stores state in memory and can crash without consequence. It may end up being driven by a saga.

I also copied the `trust_quorum` directory from the sled-agent as the `bootstore` will become responsible for the trust quorum. I haven't yet removed this code from the sled-agent, as this work is preliminary and I don't want to hook it in yet.