Check if CRDB is telling us to retry the txn
Match against DieselDatabaseErrorKind::SerializationFailure and check if
it's a case where a CRDB transaction didn't fail but needs to be
retried. This can happen if there's too much database contention, and
requires clients to retry the request.

Return HTTP 429 in this case to signal to clients that a retry is
required.